requirements.txt
paho-mqtt
dash
plotly
pandas
boto3
